pyproject.toml
[project]
name = "payments-system-django"
version = "0.1.0"
description = "Test project"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django==4.2.17",
    "pydantic>=2.11.5",
    "python-stdnum>=2.1",
]
